chore: add onceChan test
