+++ internal/graceful/graceful.go
@@ -58,27 +58,6 @@ func RegisterServerFunc(name string, serveFunc func(ctx context.Context) error, 
 	})
 }
 
-type onceChan[T any] struct {
-	ch   chan T
-	once sync.Once
-}
-
-func newOnceChan[T any]() *onceChan[T] {
-	return &onceChan[T]{
-		ch: make(chan T),
-	}
-}
-
-func (o *onceChan[T]) Close() {
-	o.once.Do(func() {
-		close(o.ch)
-	})
-}
-
-func (o *onceChan[T]) Done() <-chan T {
-	return o.ch
-}
-
 const shutdownTimeout = 10 * time.Second
 
 // Serve starts all registered servers and waits for a shutdown signal or any server to stop.
