+++ internal/graceful/oncechan_test.go
@@ -1,0 +1,10 @@
+package graceful
+
+import "testing"
+
+func TestOnceChan(_ *testing.T) {
+	ch := newOnceChan[int]()
+	ch.Close()
+	ch.Close()
+	<-ch.Done()
+}
